requirements.txt
asgiref==3.2.3
astroid==2.3.3
Click==7.0
celery==5.0.5
dj-database-url==0.5.0
Django==3.0.2
django-extensions==2.2.5
django-heroku==0.3.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-decouple==3.3
pytz==2019.3
redis==3.3.11
six==1.13.0
sqlparse==0.3.0
Unipath==1.1
whitenoise==5.0.1
wrapt==1.11.2
